split up release train
